Fix significant bits for enum. Fix more warnings (Syc).

+++ src/z80/argloc.c
@@ -466,12 +466,13 @@ z80_argloc_entry_t *z80_argloc_next(z80_argloc_entry_t *cur)
 void z80_argloc_entry_vainfo(z80_argloc_entry_t *entry, z80_vainfo_t *vainfo)
 {
 	z80ic_r16_t r16;
-	int i;
+	unsigned i;
 
 	if (entry->stack_sz > 0) {
 		/* Argument is stored on the stack */
 		assert(entry->reg_entries == 0);
-		vainfo->cur_off = 4 + entry->stack_off + entry->stack_sz;
+		vainfo->cur_off = 4 + (int16_t)(entry->stack_off +
+		    entry->stack_sz);
 		vainfo->cur_rel = z80sf_end;
 		vainfo->rem_bytes = 0;
 	} else {
@@ -488,11 +489,12 @@ void z80_argloc_entry_vainfo(z80_argloc_entry_t *entry, z80_vainfo_t *vainfo)
 		if (i + 1 < z80_r16_alloc_num) {
 			/* Still some registers left */
 			vainfo->rem_bytes = 6 - (i * 2 + 2);
-			vainfo->cur_off = 2 + i * 2;
+			vainfo->cur_off = 2 + (int16_t)i * 2;
 			vainfo->cur_rel = z80sf_begin;
 		} else {
 			/* Further arguments on the stack */
-			vainfo->cur_off = 4 + entry->stack_off + entry->stack_sz;
+			vainfo->cur_off = 4 + (int16_t)(entry->stack_off +
+			    entry->stack_sz);
 			vainfo->cur_rel = z80sf_end;
 			vainfo->rem_bytes = 0;
 		}
